+++ src/enemy.cpp
@@ -2,7 +2,7 @@
 #include <player.h>
 #include <map.h>
 
-Enemy::Enemy(float X, float Y, float w, float h, int hw, int hh,
+Enemy::Enemy(float X, float Y, int w, int h, int hw, int hh,
 			int sw, int sh, int sim, int sfm) :
     Object2D({X, Y, w, h, hw, hh}, {sw, sh, sim, sfm})
 {}
